just put the vue file into tsx file instead
