--- src/features/advanced/cx-search-bar/cx-search-bar.module.css
@@ -1,12 +1,0 @@
-.table {
-  tbody tr:nth-child(2n+1) {
-    position: relative;
-  }
-
-  tbody tr:nth-child(2n+1)::after {
-    content: "";
-    position: absolute;
-    inset: 0px;
-    pointer-events: none;
-  }
-}
