+++ src/features/advanced/index.ts
@@ -5,7 +5,7 @@ import './bbl-hide-book-value';
 import './bs-hide-zero-workforce';
 import './cogcpex-clean-labels';
 import './context-controls-no-hover';
-import './cx-search-bar/cx-search-bar';
+import './cx-search-bar';
 import './cxos-hide-exchange';
 import './cxpo-shorten-fields';
 import './finla-hide-ecd';
